fix conficts / upgrade to the newest spark authorizer
